requirements.txt


django
djangorestframework
djangorestframework_simplejwt
django-filter